add visit for switching solver

+++ src/collision/rigid/rigid.h
@@ -22,6 +22,6 @@ public:
     const glm::mat4 m_model;
     const glm::mat4 m_inverseModel;
     const glm::mat4 m_inverseTransposeModel;
-    const float kappa = 1e2;
+    const double kappa = 1e2;
     char* name = nullptr;
 };
